+++ CHANGELOG.md
@@ -20,6 +20,7 @@ In addition to original Keep-a-Changelog, we use following rules:
 ### Changed
 
 ### Fixed
+- Deseialize `Option::Some`.
 - Recursive implementation of `ruststep::tables::EntityTable::{get_owned, owned_iter}` for select types.
 
 ### Internal